Move deck list logic into feature
