+++ src/features/deck/list/index.ts
@@ -1,0 +1,6 @@
+export {
+  buildDeckListSections,
+  type DeckListItem,
+  type DeckListSections,
+  type DeckListStudyProgress,
+} from "./model/buildDeckListSections";
