+++ src/features/deck/list/model/buildDeckListSections.ts
@@ -1,12 +1,11 @@
-/**
- * @file Builds the view model for the Deck List Page.
- * Keeping these calculations outside React makes their inputs, outputs, and edge cases easier to
- * understand and test.
- */
-
 import type { Card } from "@/entities/card";
 import type { Deck, DeckId } from "@/entities/deck";
-import type { StudySession } from "@/features/study";
+
+interface DeckListStudySession {
+  cardOrderIds: string[];
+  currentIndex: number;
+  lastStudiedAt: number;
+}
 
 export interface DeckListStudyProgress {
   currentIndex: number;
@@ -25,21 +24,12 @@ export interface DeckListSections {
   other: DeckListItem[];
 }
 
-/**
- * Orders two deck-list items alphabetically by deck name.
- * The named comparison keeps the inactive-deck section stable and easy to scan.
- */
 const compareNames = (left: DeckListItem, right: DeckListItem) => left.deck.name.localeCompare(right.deck.name);
 
-/**
- * Builds deck list sections from the supplied application values.
- * The returned value is ready for the next layer, so callers do not need to repeat assembly or
- * defaulting rules.
- */
 export const buildDeckListSections = (
   decks: Deck[],
   cards: Card[],
-  sessionsByDeckId: Partial<Record<DeckId, StudySession>>
+  sessionsByDeckId: Partial<Record<DeckId, DeckListStudySession>>
 ): DeckListSections => {
   const cardCounts = new Map<DeckId, number>();
   for (const card of cards) cardCounts.set(card.deckId, (cardCounts.get(card.deckId) ?? 0) + 1);
